Added Booking model and related UI changes
In this commit, we have introduced a new `Booking` model and made several changes to the UI. The `Booking` model has been added to the `ApplicationDbContext` class, allowing the application to interact with the new `Booking` table in the database. The UI changes include modifications to the `MainWindow.xaml` file, where we have updated the `SelectionChanged` event of the `BrandSelection` ComboBox and added `Click` events to the `SearchButton` and `BookButton`. We have also added two new `DataGrid` elements to the `Laptops` and `Bookings` tabs. In the `MainWindow.xaml.cs` file, we have added several new methods to handle booking submissions, check laptop availability, handle search button clicks, and validate the search form. Lastly, we have performed a database migration using Entity Framework Core to add the `Bookings` table to the database, and updated the `ApplicationDbContextModelSnapshot.cs` file to reflect this change.